Move index tests that only need empty bucket/no indexes to `db/indextest`
