+++ db/indexes_test.go
@@ -11,7 +11,6 @@ licenses/APL2.txt.
 package db
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"log"
@@ -22,183 +21,7 @@ import (
 	"github.com/couchbase/sync_gateway/base"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"golang.org/x/sync/errgroup"
 )
-
-// TestInitializeIndexes ensures all of SG's indexes can be built using both values of xattrs, with all N1QLStore implementations
-func TestInitializeIndexes(t *testing.T) {
-	if base.TestsDisableGSI() {
-		t.Skip("This test only works with Couchbase Server and UseViews=false")
-	}
-	base.LongRunningTest(t)
-
-	tests := []struct {
-		xattrs           bool
-		collections      bool
-		clusterN1QLStore bool
-	}{
-		{true, false, false},
-		{false, false, false},
-		{true, true, false},
-		{false, true, false},
-		{true, false, true},
-		{false, false, true},
-		{true, true, true},
-		{false, true, true},
-	}
-
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("xattrs=%v collections=%v clusterN1QL=%v", test.xattrs, test.collections, test.clusterN1QLStore), func(t *testing.T) {
-			options := DatabaseContextOptions{
-				EnableXattr: test.xattrs,
-			}
-			if test.collections {
-				base.TestRequiresCollections(t)
-				// uses Scopes implicitly through test harness
-			} else {
-				options.Scopes = GetScopesOptionsDefaultCollectionOnly(t)
-			}
-			db, ctx := SetupTestDBWithOptions(t, options)
-			defer db.Close(ctx)
-			collection := GetSingleDatabaseCollection(t, db.DatabaseContext)
-
-			gocbBucket, err := base.AsGocbV2Bucket(db.Bucket)
-			require.NoError(t, err)
-
-			n1qlStore, ok := base.AsN1QLStore(collection.dataStore)
-			require.True(t, ok)
-
-			if test.clusterN1QLStore {
-				n1qlStore, err = base.NewClusterOnlyN1QLStore(gocbBucket.GetCluster(), gocbBucket.BucketName(), collection.ScopeName, collection.Name)
-				require.NoError(t, err)
-			}
-
-			// drop and restore indexes between tests, to the way the bucket pool would
-			defer func() {
-				options := InitializeIndexOptions{
-					NumReplicas: 0,
-					Serverless:  db.IsServerless(),
-					UseXattrs:   base.TestUseXattrs(),
-				}
-				if db.OnlyDefaultCollection() {
-					options.MetadataIndexes = IndexesAll
-				}
-
-				err := dropAndInitializeIndexes(base.TestCtx(t), n1qlStore, options)
-				assert.NoError(t, err)
-			}()
-
-			// add and drop indexes that may be different from the way the bucket pool expects, so use specific options here for test
-			xattrSpecificIndexOptions := InitializeIndexOptions{
-				NumReplicas: 0,
-				Serverless:  db.IsServerless(),
-				UseXattrs:   test.xattrs,
-			}
-			if db.OnlyDefaultCollection() {
-				xattrSpecificIndexOptions.MetadataIndexes = IndexesAll
-			}
-
-			// Make sure we can drop and reinitialize twice
-			for i := 0; i < 2; i++ {
-				err := dropAndInitializeIndexes(base.TestCtx(t), n1qlStore, xattrSpecificIndexOptions)
-				require.NoError(t, err, "Error dropping and initialising all indexes on bucket")
-			}
-		})
-	}
-}
-
-// TestInitializeIndexesConcurrentMultiNode simulates a multi-node SG cluster starting up and racing to create indexes.
-func TestInitializeIndexesConcurrentMultiNode(t *testing.T) {
-	if base.TestsDisableGSI() {
-		t.Skip("This test only works with Couchbase Server and UseViews=false")
-	}
-	base.LongRunningTest(t)
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-
-	// Pick a high enough number that it's likely to trigger the concurrent/race conditions we're testing for.
-	const numNodes = 10
-
-	tests := []struct {
-		numNodes         int
-		xattrs           bool
-		collections      bool
-		clusterN1QLStore bool
-	}{
-		{numNodes, true, false, false},
-		{numNodes, false, false, false},
-		{numNodes, true, true, false},
-		{numNodes, false, true, false},
-		{numNodes, true, false, true},
-		{numNodes, false, false, true},
-		{numNodes, true, true, true},
-		{numNodes, false, true, true},
-	}
-
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("xattrs=%v collections=%v clusterN1QL=%v", test.xattrs, test.collections, test.clusterN1QLStore), func(t *testing.T) {
-			options := DatabaseContextOptions{
-				EnableXattr: test.xattrs,
-			}
-			if test.collections {
-				base.TestRequiresCollections(t)
-				// uses Scopes implicitly through test harness
-			} else {
-				options.Scopes = GetScopesOptionsDefaultCollectionOnly(t)
-			}
-			db, ctx := SetupTestDBWithOptions(t, options)
-			defer db.Close(ctx)
-			collection := GetSingleDatabaseCollection(t, db.DatabaseContext)
-
-			gocbBucket, err := base.AsGocbV2Bucket(db.Bucket)
-			require.NoError(t, err)
-
-			n1qlStore, ok := base.AsN1QLStore(collection.dataStore)
-			require.True(t, ok)
-
-			if test.clusterN1QLStore {
-				n1qlStore, err = base.NewClusterOnlyN1QLStore(gocbBucket.GetCluster(), gocbBucket.BucketName(), collection.ScopeName, collection.Name)
-				require.NoError(t, err)
-			}
-
-			// drop and restore indexes between tests, to the way the bucket pool would
-			defer func() {
-				options := InitializeIndexOptions{
-					NumReplicas: 0,
-					Serverless:  db.IsServerless(),
-					UseXattrs:   base.TestUseXattrs(),
-				}
-				if db.OnlyDefaultCollection() {
-					options.MetadataIndexes = IndexesAll
-				}
-
-				err := dropAndInitializeIndexes(base.TestCtx(t), n1qlStore, options)
-				assert.NoError(t, err)
-			}()
-
-			// add and drop indexes that may be different from the way the bucket pool expects, so use specific options here for test
-			xattrSpecificIndexOptions := InitializeIndexOptions{
-				NumReplicas: 0,
-				Serverless:  db.IsServerless(),
-				UseXattrs:   test.xattrs,
-			}
-			if db.OnlyDefaultCollection() {
-				xattrSpecificIndexOptions.MetadataIndexes = IndexesAll
-			}
-
-			require.NoError(t, base.DropAllIndexes(ctx, n1qlStore))
-
-			eg := errgroup.Group{}
-			for i := 0; i < test.numNodes; i++ {
-				ctx := base.CorrelationIDLogCtx(ctx, fmt.Sprintf("test-node-%d", i))
-				eg.Go(func() error {
-					return InitializeIndexes(ctx, n1qlStore, xattrSpecificIndexOptions)
-				})
-			}
-			require.NoError(t, eg.Wait())
-		})
-	}
-}
 
 func TestPostUpgradeIndexesSimple(t *testing.T) {
 
@@ -511,25 +334,4 @@ func TestIsIndexerError(t *testing.T) {
 	assert.False(t, isIndexerError(err))
 	err = errors.New("err:[5000]  MCResponse status=KEY_ENOENT, opcode=0x89, opaque=0")
 	assert.True(t, isIndexerError(err))
-}
-
-// dropAndInitializeIndexes drops and reinitialize all sync gateway indexes
-func dropAndInitializeIndexes(ctx context.Context, n1qlStore base.N1QLStore, options InitializeIndexOptions) error {
-	dropErr := base.DropAllIndexes(ctx, n1qlStore)
-	if dropErr != nil {
-		return dropErr
-	}
-
-	initErr := InitializeIndexes(ctx, n1qlStore, options)
-	if initErr != nil {
-		return initErr
-	}
-
-	// Recreate the primary index required by the test bucket pooling framework
-	err := n1qlStore.CreatePrimaryIndex(ctx, base.PrimaryIndexName, nil)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
